+++ lib/Calibre/BookList.php
@@ -11,6 +11,7 @@ namespace SebLucas\Cops\Calibre;
 
 use SebLucas\Cops\Input\Request;
 use SebLucas\Cops\Model\Entry;
+use SebLucas\Cops\Model\EntryBook;
 use SebLucas\Cops\Model\LinkNavigation;
 use SebLucas\Cops\Pages\Page;
 use SebLucas\Cops\Pages\PageQueryResult;
@@ -19,40 +20,46 @@ class BookList extends Base
 {
     public const SQL_BOOKS_ALL = 'select {0} from books ' . Book::SQL_BOOKS_LEFT_JOIN . ' where 1=1 {1} order by books.sort ';
     public const SQL_BOOKS_BY_PUBLISHER = 'select {0} from books_publishers_link, books ' . Book::SQL_BOOKS_LEFT_JOIN . '
-    where books_publishers_link.book = books.id and publisher = ? {1} order by publisher';
+    where books_publishers_link.book = books.id and publisher = ? {1} order by books.sort';
     public const SQL_BOOKS_BY_FIRST_LETTER = 'select {0} from books ' . Book::SQL_BOOKS_LEFT_JOIN . '
     where upper (books.sort) like ? {1} order by books.sort';
+    public const SQL_BOOKS_BY_PUB_YEAR = 'select {0} from books ' . Book::SQL_BOOKS_LEFT_JOIN . '
+    where substr(date(books.pubdate), 1, 4) = ? {1} order by books.sort';
     public const SQL_BOOKS_BY_AUTHOR = 'select {0} from books_authors_link, books ' . Book::SQL_BOOKS_LEFT_JOIN . '
     left outer join books_series_link on books_series_link.book = books.id
     where books_authors_link.book = books.id and author = ? {1} order by series desc, series_index asc, pubdate asc';
     public const SQL_BOOKS_BY_SERIE = 'select {0} from books_series_link, books ' . Book::SQL_BOOKS_LEFT_JOIN . '
     where books_series_link.book = books.id and series = ? {1} order by series_index';
     public const SQL_BOOKS_BY_TAG = 'select {0} from books_tags_link, books ' . Book::SQL_BOOKS_LEFT_JOIN . '
-    where books_tags_link.book = books.id and tag = ? {1} order by sort';
+    where books_tags_link.book = books.id and tag = ? {1} order by books.sort';
     public const SQL_BOOKS_BY_LANGUAGE = 'select {0} from books_languages_link, books ' . Book::SQL_BOOKS_LEFT_JOIN . '
-    where books_languages_link.book = books.id and lang_code = ? {1} order by sort';
+    where books_languages_link.book = books.id and lang_code = ? {1} order by books.sort';
     public const SQL_BOOKS_BY_CUSTOM = 'select {0} from {2}, books ' . Book::SQL_BOOKS_LEFT_JOIN . '
-    where {2}.book = books.id and {2}.{3} = ? {1} order by sort';
+    where {2}.book = books.id and {2}.{3} = ? {1} order by books.sort';
     public const SQL_BOOKS_BY_CUSTOM_BOOL_TRUE = 'select {0} from {2}, books ' . Book::SQL_BOOKS_LEFT_JOIN . '
-    where {2}.book = books.id and {2}.value = 1 {1} order by sort';
+    where {2}.book = books.id and {2}.value = 1 {1} order by books.sort';
     public const SQL_BOOKS_BY_CUSTOM_BOOL_FALSE = 'select {0} from {2}, books ' . Book::SQL_BOOKS_LEFT_JOIN . '
-    where {2}.book = books.id and {2}.value = 0 {1} order by sort';
+    where {2}.book = books.id and {2}.value = 0 {1} order by books.sort';
     public const SQL_BOOKS_BY_CUSTOM_NULL = 'select {0} from books ' . Book::SQL_BOOKS_LEFT_JOIN . '
-    where books.id not in (select book from {2}) {1} order by sort';
+    where books.id not in (select book from {2}) {1} order by books.sort';
     public const SQL_BOOKS_BY_CUSTOM_RATING = 'select {0} from books ' . Book::SQL_BOOKS_LEFT_JOIN . '
     left join {2} on {2}.book = books.id
     left join {3} on {3}.id = {2}.{4}
-    where {3}.value = ?  order by sort';
+    where {3}.value = ?  order by books.sort';
     public const SQL_BOOKS_BY_CUSTOM_RATING_NULL = 'select {0} from books ' . Book::SQL_BOOKS_LEFT_JOIN . '
     left join {2} on {2}.book = books.id
     left join {3} on {3}.id = {2}.{4}
-    where ((books.id not in (select {2}.book from {2})) or ({3}.value = 0)) {1} order by sort';
+    where ((books.id not in (select {2}.book from {2})) or ({3}.value = 0)) {1} order by books.sort';
     public const SQL_BOOKS_BY_CUSTOM_DATE = 'select {0} from {2}, books ' . Book::SQL_BOOKS_LEFT_JOIN . '
-    where {2}.book = books.id and date({2}.value) = ? {1} order by sort';
+    where {2}.book = books.id and date({2}.value) = ? {1} order by books.sort';
+    public const SQL_BOOKS_BY_CUSTOM_YEAR = 'select {0} from {2}, books ' . Book::SQL_BOOKS_LEFT_JOIN . '
+    where {2}.book = books.id and substr(date({2}.value), 1, 4) = ? {1} order by {2}.value';
     public const SQL_BOOKS_BY_CUSTOM_DIRECT = 'select {0} from {2}, books ' . Book::SQL_BOOKS_LEFT_JOIN . '
-    where {2}.book = books.id and {2}.value = ? {1} order by sort';
+    where {2}.book = books.id and {2}.value = ? {1} order by books.sort';
+    public const SQL_BOOKS_BY_CUSTOM_RANGE = 'select {0} from {2}, books ' . Book::SQL_BOOKS_LEFT_JOIN . '
+    where {2}.book = books.id and {2}.value >= ? and {2}.value <= ? {1} order by {2}.value';
     public const SQL_BOOKS_BY_CUSTOM_DIRECT_ID = 'select {0} from {2}, books ' . Book::SQL_BOOKS_LEFT_JOIN . '
-    where {2}.book = books.id and {2}.id = ? {1} order by sort';
+    where {2}.book = books.id and {2}.id = ? {1} order by books.sort';
     public const SQL_BOOKS_QUERY = 'select {0} from books ' . Book::SQL_BOOKS_LEFT_JOIN . '
     where (
     exists (select null from authors, books_authors_link where book = books.id and author = authors.id and authors.name like ?) or
@@ -61,24 +68,44 @@ class BookList extends Base
     exists (select null from publishers, books_publishers_link where book = books.id and books_publishers_link.publisher = publishers.id and publishers.name like ?) or
     title like ?) {1} order by books.sort';
     public const SQL_BOOKS_RECENT = 'select {0} from books ' . Book::SQL_BOOKS_LEFT_JOIN . '
-    where 1=1 {1} order by timestamp desc limit ';
+    where 1=1 {1} order by books.timestamp desc limit ';
     public const SQL_BOOKS_BY_RATING = 'select {0} from books ' . Book::SQL_BOOKS_LEFT_JOIN . '
-    where books_ratings_link.book = books.id and ratings.id = ? {1} order by sort';
+    where books_ratings_link.book = books.id and ratings.id = ? {1} order by books.sort';
     public const SQL_BOOKS_BY_RATING_NULL = 'select {0} from books ' . Book::SQL_BOOKS_LEFT_JOIN . '
-    where ((books.id not in (select book from books_ratings_link)) or (ratings.rating = 0)) {1} order by sort';
+    where ((books.id not in (select book from books_ratings_link)) or (ratings.rating = 0)) {1} order by books.sort';
 
     public const BAD_SEARCH = 'QQQQQ';
 
     public Request $request;
     protected mixed $numberPerPage = null;
     protected array $ignoredCategories = [];
+    public mixed $orderBy = null;
 
     public function __construct(Request $request, mixed $database = null, mixed $numberPerPage = null)
     {
         $this->request = $request;
-        $this->databaseId = $database ?? $this->request->get('db');
+        $this->databaseId = $database ?? $this->request->get('db', null, '/^\d+$/');
         $this->numberPerPage = $numberPerPage ?? $this->request->option("max_item_per_page");
         $this->ignoredCategories = $this->request->option('ignored_categories');
+        $this->setOrderBy();
+    }
+
+    protected function setOrderBy()
+    {
+        $this->orderBy = $this->request->get('sort', null, '/^\w+(\s+(asc|desc)|)$/i');
+        //$this->orderBy ??= $this->request->option('sort');
+    }
+
+    protected function getOrderBy()
+    {
+        return match ($this->orderBy) {
+            'title' => 'books.sort',
+            'author' => 'books.author_sort',
+            'pubdate' => 'books.pubdate desc',
+            'rating' => 'ratings.rating desc',
+            'timestamp' => 'books.timestamp desc',
+            default => $this->orderBy,
+        };
     }
 
     public function getBookCount()
@@ -177,22 +204,51 @@ class BookList extends Base
     }
 
     /**
-     * @param $customColumn CustomColumn
-     * @param $id integer
-     * @param $n integer
+     * Summary of getBooksByCustom
+     * @param CustomColumnType $columnType
+     * @param integer $id
+     * @param integer $n
      * @return array
      */
-    public function getBooksByCustom($customColumn, $id, $n)
+    public function getBooksByCustom($columnType, $id, $n)
     {
-        [$query, $params] = $customColumn->getQuery($id);
+        [$query, $params] = $columnType->getQuery($id);
 
         return $this->getEntryArray($query, $params, $n);
     }
 
-    public function getBooksWithoutCustom($customColumn, $n)
+    /**
+     * Summary of getBooksByCustomYear
+     * @param CustomColumnTypeDate $columnType
+     * @param mixed $year
+     * @param mixed $n
+     * @return array
+     */
+    public function getBooksByCustomYear($columnType, $year, $n)
+    {
+        [$query, $params] = $columnType->getQueryByYear($year);
+
+        return $this->getEntryArray($query, $params, $n);
+    }
+
+    /**
+     * Summary of getBooksByCustomRange
+     * @param CustomColumnTypeInteger $columnType
+     * @param mixed $range
+     * @param mixed $n
+     * @return array
+     */
+    public function getBooksByCustomRange($columnType, $range, $n)
+    {
+        [$query, $params] = $columnType->getQueryByRange($range);
+
+        return $this->getEntryArray($query, $params, $n);
+    }
+
+    public function getBooksWithoutCustom($columnType, $n)
     {
         // use null here to reduce conflict with bool and int custom columns
-        [$query, $params] = $customColumn->getQuery(null);
+        [$query, $params] = $columnType->getQuery(null);
         return $this->getEntryArray($query, $params, $n);
     }
 
@@ -220,13 +276,23 @@ class BookList extends Base
         return $this->getEntryArray(self::SQL_BOOKS_QUERY, $critArray, $n);
     }
 
-    public function getBooks($n)
+    public function getAllBooks($n)
     {
         [$entryArray, $totalNumber] = $this->getEntryArray(self::SQL_BOOKS_ALL, [], $n);
         return [$entryArray, $totalNumber];
     }
 
-    public function getAllBooks()
+    public function getCountByFirstLetter()
+    {
+        return $this->getCountByGroup('substr(upper(books.sort), 1, 1)', Book::PAGE_LETTER, 'letter');
+    }
+
+    public function getCountByPubYear()
+    {
+        return $this->getCountByGroup('substr(date(books.pubdate), 1, 4)', Book::PAGE_YEAR, 'year');
+    }
+
+    public function getCountByGroup($groupField, $page, $label)
     {
         $filter = new Filter($this->request, [], "books", $this->databaseId);
         $filterString = $filter->getFilterString();
@@ -236,28 +302,33 @@ class BookList extends Base
         [, $result] = parent::executeQuery('select {0}
 from books
 where 1=1 {1}
-group by substr (upper (sort), 1, 1)
-order by substr (upper (sort), 1, 1)', 'substr (upper (sort), 1, 1) as title, count(*) as count', $filterString, $params, -1, $this->databaseId);
+group by groupid
+order by groupid', $groupField . ' as groupid, count(*) as count', $filterString, $params, -1, $this->databaseId);
 
         $entryArray = [];
         while ($post = $result->fetchObject()) {
             array_push($entryArray, new Entry(
-                $post->title,
-                Book::getEntryIdByLetter($post->title),
+                $post->groupid,
+                Book::PAGE_ID.':'.$label.':'.$post->groupid,
                 str_format(localize('bookword', $post->count), $post->count),
                 'text',
-                [new LinkNavigation('?page='.Book::PAGE_LETTER.'&id='. rawurlencode($post->title), null, null, $this->databaseId)],
+                [new LinkNavigation('?page='.$page.'&id='. rawurlencode($post->groupid), null, null, $this->databaseId)],
                 $this->databaseId,
-                '',
+                ucfirst($label),
                 $post->count
             ));
         }
         return $entryArray;
     }
 
-    public function getBooksByStartingLetter($letter, $n)
+    public function getBooksByFirstLetter($letter, $n)
     {
         return $this->getEntryArray(self::SQL_BOOKS_BY_FIRST_LETTER, [$letter . '%'], $n);
+    }
+
+    public function getBooksByPubYear($year, $n)
+    {
+        return $this->getEntryArray(self::SQL_BOOKS_BY_PUB_YEAR, [$year], $n);
     }
 
     public function getAllRecentBooks()
@@ -267,11 +338,26 @@ order by substr (upper (sort), 1, 1)', 'substr (upper (sort), 1, 1) as title, co
         return $entryArray;
     }
 
+    /**
+     * Summary of getEntryArray
+     * @param mixed $query
+     * @param mixed $params
+     * @param mixed $n
+     * @return array{0: EntryBook[], 1: integer}
+     */
     public function getEntryArray($query, $params, $n)
     {
         $filter = new Filter($this->request, $params, "books", $this->databaseId);
         $filterString = $filter->getFilterString();
         $params = $filter->getQueryParams();
+
+        if (isset($this->orderBy) && $this->orderBy !== Book::SQL_SORT) {
+            if (str_contains($query, 'order by')) {
+                $query = preg_replace('/\s+order\s+by\s+[\w.]+(\s+(asc|desc)|)\s*/i', ' order by ' . $this->getOrderBy() . ' ', $query);
+            } else {
+                $query .= ' order by ' . $this->getOrderBy() . ' ';
+            }
+        }
 
         /** @var integer $totalNumber */
         /** @var \PDOStatement $result */