+++ lib/Pages/PageAllAuthorsLetter.php
@@ -14,10 +14,21 @@ class PageAllAuthorsLetter extends Page
 {
     public function InitializeContent()
     {
+        $this->getEntries();
         $this->idPage = Author::getEntryIdByLetter($this->idGet);
-        $this->entryArray = Author::getAuthorsByStartingLetter($this->idGet, $this->n, $this->getDatabaseId());
-        $this->title = str_format(localize("splitByLetter.letter"), str_format(localize("authorword", count($this->entryArray)), count($this->entryArray)), $this->idGet);
+        $count = $this->totalNumber;
+        if ($count == -1) {
+            $count = count($this->entryArray);
+        }
+        $this->title = str_format(localize("splitByLetter.letter"), str_format(localize("authorword", $count), $count), $this->idGet);
         $this->parentTitle = "";  // localize("authors.title");
         $this->parentUri = "?page=".Author::PAGE_ALL;
+    }
+
+    public function getEntries()
+    {
+        $this->entryArray = Author::getEntriesByFirstLetter($this->request, $this->idGet, $this->n, $this->getDatabaseId());
+        $this->totalNumber = Author::countEntriesByFirstLetter($this->request, $this->idGet, $this->getDatabaseId());
+        $this->sorted = Author::SQL_SORT;
     }
 }