+++ lib/Pages/PageAllBooks.php
@@ -15,13 +15,23 @@ class PageAllBooks extends Page
 {
     public function InitializeContent()
     {
+        $this->getEntries();
         $this->idPage = Book::PAGE_ID;
         $this->title = localize("allbooks.title");
+    }
+
+    public function getEntries()
+    {
         $booklist = new BookList($this->request);
-        if ($this->request->option("titles_split_first_letter") == 1) {
-            $this->entryArray = $booklist->getAllBooks();
+        if ($this->request->option("titles_split_first_letter") == 1 || $this->request->get('letter')) {
+            $this->entryArray = $booklist->getCountByFirstLetter();
+            $this->sorted = "letter";
+        } elseif (!empty($this->request->option("titles_split_publication_year")) || $this->request->get('year')) {
+            $this->entryArray = $booklist->getCountByPubYear();
+            $this->sorted = "year";
         } else {
-            [$this->entryArray, $this->totalNumber] = $booklist->getBooks($this->n);
+            [$this->entryArray, $this->totalNumber] = $booklist->getAllBooks($this->n);
+            $this->sorted = $booklist->orderBy ?? Book::SQL_SORT;
         }
     }
 }