+++ lib/Pages/PageAllPublishers.php
@@ -14,8 +14,15 @@ class PageAllPublishers extends Page
 {
     public function InitializeContent()
     {
+        $this->getEntries();
         $this->idPage = Publisher::PAGE_ID;
         $this->title = localize("publishers.title");
-        $this->entryArray = Publisher::getAllPublishers($this->n, $this->getDatabaseId());
+    }
+
+    public function getEntries()
+    {
+        $this->entryArray = Publisher::getRequestEntries($this->request, $this->n, $this->getDatabaseId());
+        $this->totalNumber = Publisher::countRequestEntries($this->request, $this->getDatabaseId());
+        $this->sorted = Publisher::SQL_SORT;
     }
 }