+++ lib/Pages/PageAuthorDetail.php
@@ -16,12 +16,24 @@ class PageAuthorDetail extends Page
     public function InitializeContent()
     {
         $author = Author::getAuthorById($this->idGet, $this->getDatabaseId());
+        if ($this->request->get('filter')) {
+            $this->filterUri = '&a=' . $this->idGet;
+            $this->getFilters($author);
+        } else {
+            $this->getEntries();
+        }
         $this->idPage = $author->getEntryId();
         $this->title = $author->name;  // not by getTitle() = $author->sort here
+        $this->currentUri = $author->getUri();
         $this->parentTitle = localize("authors.title");
         $this->parentUri = $author->getParentUri();
+        //$seriesArray = $author->getSeries();
+    }
+
+    public function getEntries()
+    {
         $booklist = new BookList($this->request);
         [$this->entryArray, $this->totalNumber] = $booklist->getBooksByAuthor($this->idGet, $this->n);
-        //$seriesArray = $author->getSeries();
+        $this->sorted = $booklist->orderBy ?? "series desc";
     }
 }