+++ lib/Pages/PageTagDetail.php
@@ -16,11 +16,23 @@ class PageTagDetail extends Page
     public function InitializeContent()
     {
         $tag = Tag::getTagById($this->idGet, $this->getDatabaseId());
+        if ($this->request->get('filter')) {
+            $this->filterUri = '&t=' . $this->idGet;
+            $this->getFilters($tag);
+        } else {
+            $this->getEntries();
+        }
         $this->idPage = $tag->getEntryId();
         $this->title = $tag->getTitle();
+        $this->currentUri = $tag->getUri();
         $this->parentTitle = localize("tags.title");
         $this->parentUri = $tag->getParentUri();
+    }
+
+    public function getEntries()
+    {
         $booklist = new BookList($this->request);
         [$this->entryArray, $this->totalNumber] = $booklist->getBooksByTag($this->idGet, $this->n);
+        $this->sorted = $booklist->orderBy ?? "sort";
     }
 }